Added a qemu arm32 machine and updated benchmark
Added a new machine that is for a qemu ARM 32 emulated system. In tandem
with this went through the other recipes that were modified for the ARM
64 machine and updated them to also work with the ARM 32 machine. One
thing that I needed to do here was create kernel module patches to
disable some modules that were enabled for the STM32MP platform. Leaving
those modules in would cause the kernel to fail to build. Not entirely
sure why that is the case, but I probably should find out at some point.

Updated the benchmark image to also include some utilities for
displaying stats. Most useful if there is going to be a demo of the
benchmark being run.