Updated benchmark image and machine definitions
Updated the QEMU ST machines to include an override that is important to
ensure that the build is correctly including everything that needs to be
included. Because of this, there was some things that needed to be
removed from the modifications made to the linux-stm32mp recipe.

Updated some of the dependencies of the benchmark to include what it
needs.