Updated dependencies and added debugging
Updated the layer to ensure that there is a dependency on the st layer
that we are trying to augment.

Added a new configuration option to easily enable the debugging output
on the kernel. This should continue and there should be more
configurations added throughout to enable much more debugging if
possible.